@@ -11,49 +11,66 @@
 # J1 Template is licensed under the MIT License.
 # See: https://github.com/jekyll-one-org/J1 Template/blob/master/LICENSE
 # ------------------------------------------------------------------------------
+
+# ------------------------------------------------------------------------------
+# NOTE:
+# To install all gem needed for Jekyll and J1 Template:
+#   bundle install
 #
-#  To install all gem needed for Jekyll and J1 Template
-#
-#  bundle install
-#
-#  Note:  If all packages needed are installed, a list of all gem
-#         and dependencies installed for the bundle can created
-#         by running:
-#
-#         bundle list
+# TIP:
+# If all packages needed are installed, a list of all gem and dependencies
+# installed for the bundle canbe created by running:
+#   bundle list
 #
 # ------------------------------------------------------------------------------
-#
-#  If you see warnings like:
-#
+# NOTE:
+# If you see warnings like:
 #   WARN: Unresolved specs during Gem::Specification.reset
-#         cleanup your bundle by running:
-#
-#         gem cleanup
+# you may need to cleanup your bundle by running:
+#   gem cleanup
 #
 # ------------------------------------------------------------------------------
+
+# ------------------------------------------------------------------------------
+# Define the (download) source, Ruby GEMs are to be loaded from REMOTE
+#
 source 'https://rubygems.org'
 
 # ------------------------------------------------------------------------------
+# Specify your Ruby version if the J1 Project is used as an container-based
+# web application. This makes sure to use identical Ruby runtime environments
+# for BUILD and RUN (e.g. for Docker images or a Dyno on Heroku).
+#
+# ruby '3.1'
+
+# ------------------------------------------------------------------------------
 # Jekyll
-# NOTE: J1 Template is using Jekyll v3.8 and above
+# ------------------------------------------------------------------------------
+# NOTE:
+# J1 Template is using Jekyll v4.0 and above
 #
+# ------------------------------------------------------------------------------
+# NOTE:
+# Use|Install a Jekyll version loaded from 'Github' (branch: master):
+#  gem 'jekyll', github: 'jekyll/jekyll'
+# ------------------------------------------------------------------------------
+# NOTE:
+# For default, the Jekyll GEM is loaded from REMOTE
+# ------------------------------------------------------------------------------
+#
+gem 'jekyll', '~> 4.0'
 
 # ------------------------------------------------------------------------------
-# Use Jekyll version from GH master
+# Install Webrick GEM (internally used Web Server) if Ruby V3 is used
 #
-# Support for Ruby version 2.7 (DEC 2019)
-# See: https://github.com/jekyll/jekyll/issues/8049
-# gem 'jekyll', github: 'jekyll/jekyll'
+install_if -> { RUBY_VERSION =~ /3/ } do
+  gem 'webrick', '~> 1.7'
+end
 
 # ------------------------------------------------------------------------------
-# Use Jekyll version from RubyGems
+# Specify the THEME GEM used for the project
 #
-gem 'jekyll', '~> 4.2'
-
-# Theme Rubies, default: J1 Template (NOT used for the development system)
-#
-gem 'j1-template', '~> 2022.4.10'
+gem 'j1-template', '~> 2022.5.0'
 
 # ------------------------------------------------------------------------------
 # PRODUCTION: Gem needed for the Jekyll and J1 prod environment
